+++ landstalker/src/3d_maps/Doors.cpp
@@ -55,7 +55,7 @@ std::pair< std::vector<uint8_t>, std::vector<uint8_t>> Doors::GetData(int roomco
 		if (m_doors.count(i) > 0)
 		{
 			offsets.push_back(static_cast<uint8_t>(lastsz + 1));
-			lastsz = m_doors.at(i).size() * 2 + 1;
+			lastsz = static_cast<int>(m_doors.at(i).size()) * 2 + 1;
 			for (const auto& d : m_doors.at(i))
 			{
 				auto b = d.GetBytes();
